Move handle_shortcut_select into page
