+++ src/ui/images/pages/mod.rs
@@ -10,6 +10,8 @@ pub use change_grid_db_id::handle_grid_change;
 pub use change_grid_db_id::handle_correct_grid_request;
 
 pub use shortcut_images_overview::render_page_shortcut_images_overview;
+pub use shortcut_images_overview::handle_shortcut_selected;
+
 pub use steam_images_overview::render_page_steam_images_overview;
 pub use select_image_type::render_page_shortcut_select_image_type;
 pub use pick_new_image::render_page_pick_image;
